If Project Honey Pot couldn't be reached the IP could be wrong.
Under certain server configurations, when Project Honey Pot could not be
reached the IP to be checked was incorrect and would be qualified as
spam.

Quote from the HTTPBL Api documentation: If the first octet in the
response is not 127 it means an error condition has occurred and your
query may not have been formatted correctly.

Reference: http://www.projecthoneypot.org/httpbl_api.php